+++ src/components/app-header/app-header.js
@@ -26,8 +26,9 @@ export default function AppHeader() {
                     </p>
                 </Link> 
             </nav>
-
-            <Logo />
+            <Link to='/'>    
+                <Logo />
+            </Link>
             <Link to='/profile' className={`${styles.link} mb-4 mt-4 pl-5 pr-5 pb-5 pt-5`}>
                 <ProfileIcon type={`${isProfile ? 'primary' : 'secondary'}`} />
                 <p className={`text text_type_main-default ${!isProfile && 'text_color_inactive'} ml-2`}>
