+++ src/components/burger-constructor/burger-constructor.js
@@ -3,14 +3,9 @@ import styles from './burger-constructor.module.css';
 import { ConstructorElement, CurrencyIcon, Button } from '@ya.praktikum/react-developer-burger-ui-components';
 import { useDispatch, useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
-import { url } from '../../utils/constants';
-import { 
-  updateIngredients, 
-  orderFetchData, 
-  addOrderIds, 
-  addIngredient, 
-  addBun
-} from '../../services/actions';
+import { orderFetchData, addOrderIds } from '../../services/actions/order-actions';
+import { addBun } from '../../services/actions/bun-actions';
+import { addIngredient, updateIngredients } from '../../services/actions/constructor-actions';
 import { useDrop } from "react-dnd";
 import update from 'immutability-helper';
 import ConstructorCard from '../constructor-card/constructor-card';
@@ -38,6 +33,7 @@ export default function BurgerConstructor() {
   const dataIngs = useSelector(store => store.constructorReducer.ingredientsInConstructor);
   const bunBurger = useSelector(store => store.bunReducer.bunInConstructor);
   const orderIds = useSelector(store => store.orderReducer.orderIds);
+  const orderReq = useSelector(store => store.orderReducer.orderReq);
   const [totalPrice, setTotalPrice] = useState(0);
 
   useEffect(() => {
@@ -63,7 +59,7 @@ export default function BurgerConstructor() {
 
   const isOpenModal = () => {
     name 
-    ? dispatch(orderFetchData(url, orderIds))
+    ? dispatch(orderFetchData(orderIds))
     : history.push('/login')
   }
 
@@ -128,15 +124,19 @@ export default function BurgerConstructor() {
       <div className={`${styles.info} mt-10`}>
         <p className="text text_type_digits-medium mr-2">{totalPrice}</p>
         <CurrencyIcon type="primary"/>
-        {bunBurger._id ?
-          <div onClick={isOpenModal}>
-            <Button type="primary" size="medium">
-              Оформить заказ
-            </Button>
-          </div>
-          :
-          <p className={`text text_type_main-default ml-5 ${styles.choice}`}>Выберите булку, чтобы сделать заказ</p>
-        }
+          {bunBurger._id ?
+            <div onClick={isOpenModal}>
+              {orderReq ?
+                <p className={`text text_type_main-default ml-5 ${styles.choice}`}>Идёт загрузка заказа...</p>
+              : 
+              <Button type="primary" size="medium">
+                Оформить заказ
+              </Button>
+              }
+            </div>
+            :
+            <p className={`text text_type_main-default ml-5 ${styles.choice}`}>Выберите булку, чтобы сделать заказ</p>
+          }
       </div>
 
     </section>
