+++ src/services/actions/action-types.js
@@ -1,0 +1,50 @@
+export const ADD_BUN = 'ADD_BUN';
+
+export const ADD_INGREDIENT = 'ADD_INGREDIENT';
+export const DELETE_INGREDIENT = 'DELETE_INGREDIENT';
+export const UPDATE_INGREDIENTS = 'UPDATE_INGREDIENTS';
+
+export const GET_INGREDIENTS_REQUEST = 'GET_INGREDIENTS_REQ';
+export const GET_INGREDIENTS_SUCCESS = 'GET_INGREDIENTS_SUCCESS';
+export const GET_INGREDIENTS_ERROR = 'GET_INGREDIENTS_ERROR';
+
+export const ADD_CART_MODAL = 'ADD_CART_MODAL';
+export const DELETE_CART_MODAL = 'DELETE_CART_MODAL';
+
+export const GET_ORDER_REQUEST = 'GET_ORDER_REQUEST';
+export const GET_ORDER_SUCCESS = 'GET_ORDER_SUCCESS';
+export const GET_ORDER_ERROR = 'GET_ORDER_ERROR';
+export const DELETE_ORDER_MODAL = 'DELETE_ORDER_MODAL';
+export const ADD_ORDER_IDS = 'ADD_ORDER_IDS';
+
+export const GET_LOGIN_REQUEST = 'GET_LOGIN_REQUEST';
+export const GET_LOGIN_SUCCESS = 'GET_LOGIN_SUCCESS';
+export const GET_LOGIN_ERROR = 'GET_LOGIN_ERROR';
+
+export const GET_REGISTER_REQUEST = 'GET_REGISTER_REQUEST';
+export const GET_REGISTER_SUCCESS = 'GET_REGISTER_SUCCESS';
+export const GET_REGISTER_ERROR = 'GET_REGISTER_ERROR';
+
+export const GET_TOKEN_REQUEST = 'GET_TOKEN_REQUEST';
+export const GET_TOKEN_SUCCESS = 'GET_TOKEN_SUCCESS';
+export const GET_TOKEN_ERROR = 'GET_TOKEN_ERROR';
+
+export const GET_FORGOT_REQUEST = 'GET_FORGOT_REQUEST';
+export const GET_FORGOT_SUCCESS = 'GET_FORGOT_SUCCESS';
+export const GET_FORGOT_ERROR = 'GET_FORGOT_ERROR';
+
+export const GET_RESET_PASSWORD_REQUEST = 'GET_RESET_PASSWORD_REQUEST';
+export const GET_RESET_PASSWORD_SUCCESS = 'GET_RESET_PASSWORD_SUCCESS';
+export const GET_RESET_PASSWORD_ERROR = 'GET_RESET_PASSWORD_ERROR';
+
+export const GET_LOGOUT_REQUEST = 'GET_LOGOUT_REQUEST';
+export const GET_LOGOUT_SUCCESS = 'GET_LOGOUT_SUCCESS';
+export const GET_LOGOUT_ERROR = 'GET_LOGOUT_ERROR';
+
+export const GET_USER_REQUEST = 'GET_USER_REQUEST';
+export const GET_USER_SUCCESS = 'GET_USER_SUCCESS';
+export const GET_USER_ERROR = 'GET_USER_ERROR';
+
+export const GET_UPD_USER_REQUEST = 'GET_UPD_USER_REQUEST';
+export const GET_UPD_USER_SUCCESS = 'GET_UPD_USER_SUCCESS';
+export const GET_UPD_USER_ERROR = 'GET_UPD_USER_ERROR';
